avoid using '1ul' as a 'size_t' literal
This code was written assuming `unsigned long` is the same width as `size_t` on
all platforms. This is not true for x86-64 Windows, where
`sizeof(unsigned long) == 4` and `sizeof(size_t) == 8`. The compiler observes:

  lib\cgraph\refstr.c(293,32): warning C4334: '<<': result of 32-bit shift
    implicitly converted to 64 bits (was 64-bit shift intended?)

Reported-by: MSVC 19.40.33811.0

+++ lib/common/utils.c
@@ -601,7 +601,7 @@ boxf
 polyBB (polygon_t* poly)
 {
     const size_t sides = poly->sides;
-    const size_t peris = MAX(poly->peripheries, 1ul);
+    const size_t peris = MAX(poly->peripheries, (size_t)1);
     pointf* verts = poly->vertices + (peris-1)*sides;
     boxf bb;
